pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-customer-support-triage"
version = "0.1.0"
description = "Real-world customer support ticket triage environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=1.54.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "customer_support_triage.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "customer_support_triage",
    "customer_support_triage.server",
]

[tool.setuptools.package-dir]
customer_support_triage = "."
"customer_support_triage.server" = "server"
